refactor: rename security -> auth0
