+++ internal/auth0/service.go
@@ -1,4 +1,4 @@
-package security
+package auth0
 
 func NewService(cfg *Config) *Service {
 	return &Service{
