+++ internal/graph/login.resolvers.go
@@ -9,11 +9,11 @@ import (
 	"fmt"
 
 	"edholm.dev/go-logging"
+	"github.com/filmstund/filmstund/internal/auth0/principal"
 	"github.com/filmstund/filmstund/internal/database/sqlc"
 	"github.com/filmstund/filmstund/internal/graph/gql"
 	"github.com/filmstund/filmstund/internal/graph/mappers"
 	"github.com/filmstund/filmstund/internal/graph/model"
-	"github.com/filmstund/filmstund/internal/security/principal"
 )
 
 func (r *mutationResolver) LoginUser(ctx context.Context) (*model.User, error) {
